requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cmdstanpy==1.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
docopt==0.6.2
fonttools==4.60.1
google-ai-generativelanguage==0.6.15
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
gspread==6.2.1
gspread-dataframe==4.0.0
h11==0.16.0
holidays==0.84
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
jiter==0.11.1
joblib==1.5.2
Js2Py==0.74
kiwisolver==1.4.9
lxml==5.3.0
matplotlib==3.10.7
nltk==3.9.2
numpy==2.3.4
oauthlib==3.3.1
openai==2.7.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
praw==7.8.1
prawcore==2.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pyjsparser==2.7.1
pyparsing==3.2.5
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2025.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stanio==0.5.1
threadpoolctl==3.6.0
tqdm==4.67.1
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.9.0

linkedin-api==2.2.1

# Backend API Dependencies
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery==5.4.0
redis==5.2.1
alembic==1.14.0
